ショートカットキーの不具合を修正。
