+++ source/WinCap/Views/MainWindow.xaml
@@ -1,0 +1,11 @@
+<Window x:Class="WinCap.Views.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:behaviors="clr-namespace:WinCap.Views.Behaviors"
+        mc:Ignorable="d"
+        ShowInTaskbar="False"
+        WindowState="Minimized">
+</Window>
