+++ source/WinCap/Views/MainWindow.xaml.cs
@@ -1,0 +1,14 @@
+namespace WinCap.Views
+{
+    /// <summary>
+    /// MainWindow.xaml の相互作用ロジック
+    /// </summary>
+    public partial class MainWindow
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+            this.Loaded += (sender, e) => this.Hide();
+        }
+    }
+}
